requirements.txt
fastapi
google-cloud-bigquery
google-cloud-firestore
